refactor: split games_provider into concern-based library parts (#3878)
Decompose session Riverpod providers into current-game, work-target, and
diplomacy part files without changing the public import path.

Refs #3878
