+++ SPEC/program/repo-and-packages.md
@@ -25,6 +25,7 @@ Monorepo layout:
 - **`core/services/observe/`** — [`ObserveModeSessionHandler`](../../app/lib/core/services/observe/observe_mode_session_handler.dart) applies observe-mode session mutations without importing feature panels.
 - **`core/services/event_bridge/`** — logic-to-app bus forwarding ([`game_event_bridge.dart`](../../app/lib/core/services/event_bridge/game_event_bridge.dart)).
 - **`core/services/region_map/`** — narrow Flame region-map widget binding surface ([`region_map_widget_bindings.dart`](../../app/lib/core/services/region_map/region_map_widget_bindings.dart)) consumed by shared `CtRegionMap` widget ([`ct_region_map.dart`](../../app/lib/widgets/ct_region_map.dart) library parts: `ct_region_map_state.dart` lifecycle/bus/tap handlers, `ct_region_map_viewport.dart` keyboard/scroll/hover/pinch input chrome).
+- **`providers/games_provider.dart`** — session game state Riverpod surface: [`games_provider_current_game.dart`](../../app/lib/providers/games_provider_current_game.dart) (loaded game, orders, intro-shown ids, main-menu autosave probe), [`games_provider_work_targets.dart`](../../app/lib/providers/games_provider_work_targets.dart) (civilian work-target availability + dev-exclusive tile reservations), [`games_provider_diplomacy.dart`](../../app/lib/providers/games_provider_diplomacy.dart) (pending diplomacy gates + tribe first-contact herald queue). Public import path unchanged (`package:colonizethis_app/providers/games_provider.dart`).
 - **`colonizethis_app_e2e_support`** (`packages/colonizethis_app_e2e_support`) hosts E2E/widget mirror helpers (`e2e_helpers.dart`, `e2e_test_shared*.dart`, panel expected-line fixtures) previously under `app/integration_test/` and `app/test/e2e_*`. `app/test` retains only barrel/contract tests (≤10 `e2e_*.dart` files); integration scenarios import the support package. Seed-42 demo fixtures remain in `app/lib/test_support/` (production/widgetbook consumers).
 - **`features/shell/new_game_leader_selection_dialog.dart`** — DLG10001 new-game leader picker dialog library: `_NewGameLeaderSelectionDialogStateBase` (shared fields + validation), `_NewGameLeaderSelectionDialogSlots` (slot label + text styles), `_NewGameLeaderSelectionDialogSlotRow` (per-slot nation/leader/profile row + responsive pickers), `_NewGameLeaderSelectionDialogSetupFields` (seed/advanced-start/infinite/terrain/footer). Public widget: `NewGameLeaderSelectionDialog`.
 - **`features/shell/new_game_setup_flow.dart`** — SHEL30001 new-game initializing flow (`runNewGameSetupAfterLeaderPick`) with progress/error dialog `part` ([new_game_setup_flow_dialogs.dart](../../app/lib/features/shell/new_game_setup_flow_dialogs.dart): `NewGameSetupProgressView`, `NewGameErrorCard`, async progress dialog host).
