+++ app/lib/providers/games_provider.dart
@@ -8,253 +8,6 @@ import '../config/constants.dart';
 import '../features/game/widgets/shell/shell_player_context.dart';
 import 'game_service_provider.dart';
 
-/// List of saved game ids. Refreshed by reading from GameService.
-final gameListIdsProvider = FutureProvider<List<String>>((ref) async {
-  final service = ref.watch(gameServiceProvider);
-  return service.listGameIds();
-});
-
-/// Currently loaded game, if any. Updated on load and after next turn.
-class CurrentGameNotifier extends Notifier<Game?> {
-  CurrentGameNotifier([this._initial]);
-
-  final Game? _initial;
-
-  @override
-  Game? build() => _initial;
-
-  void setGame(Game? game) {
-    state = game;
-  }
-
-  void clear() {
-    state = null;
-  }
-}
-
-final currentGameProvider = NotifierProvider<CurrentGameNotifier, Game?>(
-  CurrentGameNotifier.new,
-);
-
-/// True when the auto-save slot is valid. Rebuilds when [currentGameProvider] changes
-/// so the main menu updates immediately after exiting to menu. SPEC/ui/main-menu.md.
-final mainMenuAutoSaveAvailableProvider = Provider<bool>((ref) {
-  ref.watch(currentGameProvider);
-  if (!Hive.isBoxOpen(HiveBoxNames.games)) {
-    return false;
-  }
-  final service = ref.watch(gameServiceProvider);
-  return service.hasValidAutoSave();
-});
-
-/// Current-turn orders for the human player (work orders, move orders, etc.).
-/// Updated when the player assigns/cancels work in the civilian panel; passed to nextTurn and reset after resolution.
-class CurrentOrdersNotifier extends Notifier<Orders> {
-  CurrentOrdersNotifier([this._initial = const Orders()]);
-
-  final Orders _initial;
-
-  @override
-  Orders build() => _initial;
-
-  void replaceAll(Orders next) {
-    state = next;
-  }
-
-  void clear() {
-    state = const Orders();
-  }
-}
-
-final currentOrdersProvider = NotifierProvider<CurrentOrdersNotifier, Orders>(
-  CurrentOrdersNotifier.new,
-);
-
-/// Sorted work target ids for one civilian unit that have ≥1 valid tile
-/// (selected-unit availability). SPEC/program/order-suggestions.md (Refs #2133).
-///
-/// The app must not use broad [suggestWorkOrders] for per-unit Assign hot paths.
-final availableWorkTargetIdsForUnitProvider =
-    Provider.family<List<String>, String>((ref, unitId) {
-      final game = ref.watch(currentGameProvider);
-      if (game == null) return const [];
-
-      final orders = ref.watch(currentOrdersProvider);
-      final service = ref.watch(gameServiceProvider);
-      final mapData = service.getMapData(game.id);
-      final topology = mapData?.combinedTopology ?? const MapTopology();
-
-      final shell = ref.read(shellPlayerContextProvider);
-      final humanPlayerId =
-          shell.panelPlayerId ?? resolveShellPanelPlayerId(shell, game);
-      final view = buildPlayerView(game, topology, humanPlayerId);
-
-      return getAvailableWorkTargetsForUnit(
-        view: view,
-        game: game,
-        topology: topology,
-        currentOrders: orders,
-        unitId: unitId,
-        tileMapByRegion: mapData?.tileMapByRegion,
-      ).availableWorkTargetIdsSorted();
-    });
-
-/// Tile keys reserved for the human player’s Builder/Engineer/Merchant
-/// exclusivity (in-progress work + pending dev-exclusive work orders).
-///
-/// **SPEC/program/order-suggestions.md** § Dev-exclusive tile reservations.
-/// Exposed for UI/diagnostics; availability for assignment still flows from
-/// [availableWorkTargetIdsForUnitProvider] and
-/// [getValidWorkOrderTileKeysWithVisibility].
-final devExclusiveReservedWorkTileKeysProvider = Provider<Set<String>>((ref) {
-  final game = ref.watch(currentGameProvider);
-  if (game == null) return {};
-
-  final orders = ref.watch(currentOrdersProvider);
-  final shell = ref.read(shellPlayerContextProvider);
-  final humanPlayerId =
-      shell.panelPlayerId ?? resolveShellPanelPlayerId(shell, game);
-  return devExclusiveReservedTileKeysForPlayer(game, orders, humanPlayerId);
-});
-
-/// Set of game ids for which the game-start intro dialogue has been shown.
-/// SPEC/ai/dialogue-management.md § First dialogue emission point.
-class GameIdsWithIntroShownNotifier extends Notifier<Set<String>> {
-  GameIdsWithIntroShownNotifier([this._initial = const <String>{}]);
-
-  final Set<String> _initial;
-
-  @override
-  Set<String> build() => _initial;
-
-  void markShown(String gameId) {
-    state = {...state, gameId};
-  }
-}
-
-final gameIdsWithIntroShownProvider =
-    NotifierProvider<GameIdsWithIntroShownNotifier, Set<String>>(
-      GameIdsWithIntroShownNotifier.new,
-    );
-
-/// At most one blocking diplomacy gate from turn resolution (overture, intervention, CTA).
-/// SPEC/ui/pending-diplomacy-state.md, SPEC/program/dialogue-system.md.
-sealed class PendingDiplomacyState {
-  const PendingDiplomacyState();
-}
-
-final class PendingDiplomacyOvertures extends PendingDiplomacyState {
-  const PendingDiplomacyOvertures(this.offers);
-  final List<OvertureOffer> offers;
-}
-
-final class PendingDiplomacyIntervention extends PendingDiplomacyState {
-  const PendingDiplomacyIntervention(this.prompts);
-  final List<InterventionPrompt> prompts;
-}
-
-final class PendingDiplomacyCallToArms extends PendingDiplomacyState {
-  const PendingDiplomacyCallToArms(this.pending);
-  final List<CallToArmsPending> pending;
-}
-
-final class PendingDiplomacyFtp extends PendingDiplomacyState {
-  const PendingDiplomacyFtp(this.offers);
-  final List<FtpOffer> offers;
-}
-
-class PendingDiplomacyNotifier extends Notifier<PendingDiplomacyState?> {
-  PendingDiplomacyNotifier([this._initial]);
-
-  final PendingDiplomacyState? _initial;
-
-  @override
-  PendingDiplomacyState? build() => _initial;
-
-  void setOvertures(List<OvertureOffer> offers) {
-    state = PendingDiplomacyOvertures(offers);
-  }
-
-  void setIntervention(List<InterventionPrompt> prompts) {
-    state = PendingDiplomacyIntervention(prompts);
-  }
-
-  void setCallToArms(List<CallToArmsPending> pending) {
-    state = PendingDiplomacyCallToArms(pending);
-  }
-
-  void setFtp(List<FtpOffer> offers) {
-    state = PendingDiplomacyFtp(offers);
-  }
-
-  void clear() {
-    state = null;
-  }
-}
-
-final pendingDiplomacyProvider =
-    NotifierProvider<PendingDiplomacyNotifier, PendingDiplomacyState?>(
-      PendingDiplomacyNotifier.new,
-    );
-
-/// Payload for one tribe first-contact herald (OVL80001).
-class TribeFirstContactHeraldPayload {
-  const TribeFirstContactHeraldPayload({
-    required this.tribeId,
-    required this.tribeName,
-    required this.capitalName,
-  });
-
-  final String tribeId;
-  final String tribeName;
-  final String capitalName;
-}
-
-/// Herald keys `"$gameId|$tribeId"` already shown this session (OVL80001).
-class TribeFirstContactHeraldsShownNotifier extends Notifier<Set<String>> {
-  TribeFirstContactHeraldsShownNotifier([this._initial = const <String>{}]);
-
-  final Set<String> _initial;
-
-  @override
-  Set<String> build() => _initial;
-
-  bool isShown(String gameId, String tribeId) =>
-      state.contains('$gameId|$tribeId');
-
-  void markShown(String gameId, String tribeId) {
-    state = {...state, '$gameId|$tribeId'};
-  }
-}
-
-final tribeFirstContactHeraldsShownProvider =
-    NotifierProvider<TribeFirstContactHeraldsShownNotifier, Set<String>>(
-      TribeFirstContactHeraldsShownNotifier.new,
-    );
-
-/// FIFO queue of tribe first-contact heralds pending presentation.
-class TribeFirstContactHeraldQueueNotifier
-    extends Notifier<List<TribeFirstContactHeraldPayload>> {
-  TribeFirstContactHeraldQueueNotifier([this._initial = const []]);
-
-  final List<TribeFirstContactHeraldPayload> _initial;
-
-  @override
-  List<TribeFirstContactHeraldPayload> build() => List.of(_initial);
-
-  void enqueue(TribeFirstContactHeraldPayload payload) {
-    if (state.any((p) => p.tribeId == payload.tribeId)) return;
-    state = [...state, payload];
-  }
-
-  void dequeueHead() {
-    if (state.isEmpty) return;
-    state = state.sublist(1);
-  }
-}
-
-final tribeFirstContactHeraldQueueProvider =
-    NotifierProvider<
-      TribeFirstContactHeraldQueueNotifier,
-      List<TribeFirstContactHeraldPayload>
-    >(TribeFirstContactHeraldQueueNotifier.new);
+part 'games_provider_current_game.dart';
+part 'games_provider_work_targets.dart';
+part 'games_provider_diplomacy.dart';
